split penumbra-wasm and penumbra-extension relevant code
