+++ crates/penumbra-extension/Cargo.toml
@@ -1,8 +1,11 @@
 [package]
-name = "penumbra-wasm"
-version = {workspace = true}
-authors = {workspace = true}
-edition = {workspace = true}
+name = "penumbra-extension"
+authors.workspace = true
+edition.workspace = true
+version.workspace = true
+repository.workspace = true
+homepage.workspace = true
+license.workspace = true
 
 [lib]
 crate-type = ["cdylib", "rlib"]
@@ -27,7 +30,6 @@ penumbra-shielded-pool = {workspace = true, default-features = false}
 penumbra-stake = {workspace = true, default-features = false}
 penumbra-tct = {workspace = true, default-features = true}
 penumbra-transaction = {workspace = true, default-features = false}
-penumbra-extension = { path = "../penumbra-extension", default-features = false }
 anyhow = {workspace = true}
 ark-ff = {workspace = true, features = ["std"], default-features = true}
 base64 = {workspace = true}
