+++ crates/penumbra-extension/src/error.rs
