+++ crates/penumbra-extension/src/lib.rs
@@ -2,13 +2,11 @@
 // Requires nightly.
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
 
-extern crate core;
-
-pub mod build;
-pub mod dex;
-// pub mod error;
-pub mod keys;
+pub mod error;
+pub mod note_record;
+pub mod planner;
+pub mod storage;
+pub mod swap_record;
 pub mod tx;
 pub mod utils;
 pub mod view_server;
-pub mod wasm_planner;
