+++ crates/penumbra-extension/src/note_record.rs
