+++ crates/penumbra-extension/src/storage.rs
