+++ crates/penumbra-extension/src/view_server.rs
@@ -1,0 +1,56 @@
+use crate::note_record::SpendableNoteRecord;
+use crate::swap_record::SwapRecord;
+use penumbra_tct as tct;
+use serde::{Deserialize, Serialize};
+use tct::storage::{StoreCommitment, StoreHash, StoredPosition, Updates};
+use tct::{Forgotten, Tree};
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct StoredTree {
+    last_position: Option<StoredPosition>,
+    last_forgotten: Option<Forgotten>,
+    hashes: Vec<StoreHash>,
+    commitments: Vec<StoreCommitment>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct ScanBlockResult {
+    pub height: u64,
+    pub sct_updates: Updates,
+    pub new_notes: Vec<SpendableNoteRecord>,
+    pub new_swaps: Vec<SwapRecord>,
+}
+
+impl ScanBlockResult {
+    pub fn new(
+        height: u64,
+        sct_updates: Updates,
+        new_notes: Vec<SpendableNoteRecord>,
+        new_swaps: Vec<SwapRecord>,
+    ) -> ScanBlockResult {
+        Self {
+            height,
+            sct_updates,
+            new_notes,
+            new_swaps,
+        }
+    }
+}
+
+pub fn load_tree(stored_tree: StoredTree) -> Tree {
+    let stored_position: StoredPosition = stored_tree.last_position.unwrap_or_default();
+    let mut add_commitments = Tree::load(
+        stored_position,
+        stored_tree.last_forgotten.unwrap_or_default(),
+    );
+
+    for store_commitment in &stored_tree.commitments {
+        add_commitments.insert(store_commitment.position, store_commitment.commitment)
+    }
+    let mut add_hashes = add_commitments.load_hashes();
+
+    for stored_hash in &stored_tree.hashes {
+        add_hashes.insert(stored_hash.position, stored_hash.height, stored_hash.hash);
+    }
+    add_hashes.finish()
+}
