+++ crates/wasm/src/utils.rs
@@ -3,19 +3,8 @@ use wasm_bindgen::JsValue;
 
 use penumbra_proto::DomainType;
 
-use crate::error::WasmResult;
-use crate::utils;
-
-pub fn set_panic_hook() {
-    // When the `console_error_panic_hook` feature is enabled, we can call the
-    // `set_panic_hook` function at least once during initialization, and then
-    // we will get better error messages if our code ever panics.
-    //
-    // For more details see
-    // https://github.com/rustwasm/console_error_panic_hook#readme
-    #[cfg(feature = "console_error_panic_hook")]
-    console_error_panic_hook::set_once();
-}
+use penumbra_extension::error::WasmResult;
+use penumbra_extension::utils;
 
 /// decode SCT root
 /// Arguments:
