+++ crates/wasm/src/view_server.rs
@@ -1,7 +1,7 @@
 use std::convert::TryInto;
 use std::{collections::BTreeMap, str::FromStr};
 
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use serde_wasm_bindgen::Error;
 use serde_wasm_bindgen::Serializer;
 use wasm_bindgen::prelude::wasm_bindgen;
@@ -11,51 +11,20 @@ use penumbra_asset::asset::{Id, Metadata};
 use penumbra_compact_block::{CompactBlock, StatePayload};
 use penumbra_dex::lp::position::Position;
 use penumbra_dex::lp::LpNft;
+use penumbra_extension::view_server::{load_tree, ScanBlockResult, StoredTree};
 use penumbra_keys::FullViewingKey;
 use penumbra_sct::Nullifier;
 use penumbra_shielded_pool::note;
 use penumbra_tct as tct;
 use penumbra_tct::Witness::*;
-use tct::storage::{StoreCommitment, StoreHash, StoredPosition, Updates};
+use tct::storage::{StoredPosition, Updates};
 use tct::{Forgotten, Tree};
 
-use crate::error::WasmResult;
-use crate::note_record::SpendableNoteRecord;
-use crate::storage::IndexedDBStorage;
-use crate::swap_record::SwapRecord;
-use crate::utils;
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct StoredTree {
-    last_position: Option<StoredPosition>,
-    last_forgotten: Option<Forgotten>,
-    hashes: Vec<StoreHash>,
-    commitments: Vec<StoreCommitment>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct ScanBlockResult {
-    height: u64,
-    sct_updates: Updates,
-    new_notes: Vec<SpendableNoteRecord>,
-    new_swaps: Vec<SwapRecord>,
-}
-
-impl ScanBlockResult {
-    pub fn new(
-        height: u64,
-        sct_updates: Updates,
-        new_notes: Vec<SpendableNoteRecord>,
-        new_swaps: Vec<SwapRecord>,
-    ) -> ScanBlockResult {
-        Self {
-            height,
-            sct_updates,
-            new_notes,
-            new_swaps,
-        }
-    }
-}
+use penumbra_extension::error::WasmResult;
+use penumbra_extension::note_record::SpendableNoteRecord;
+use penumbra_extension::storage::IndexedDBStorage;
+use penumbra_extension::swap_record::SwapRecord;
+use penumbra_extension::utils;
 
 #[wasm_bindgen]
 pub struct ViewServer {
@@ -324,22 +293,4 @@ impl ViewServer {
         let denom = lp_nft.denom();
         serde_wasm_bindgen::to_value(&denom)
     }
-}
-
-pub fn load_tree(stored_tree: StoredTree) -> Tree {
-    let stored_position: StoredPosition = stored_tree.last_position.unwrap_or_default();
-    let mut add_commitments = Tree::load(
-        stored_position,
-        stored_tree.last_forgotten.unwrap_or_default(),
-    );
-
-    for store_commitment in &stored_tree.commitments {
-        add_commitments.insert(store_commitment.position, store_commitment.commitment)
-    }
-    let mut add_hashes = add_commitments.load_hashes();
-
-    for stored_hash in &stored_tree.hashes {
-        add_hashes.insert(stored_hash.position, stored_hash.height, stored_hash.hash);
-    }
-    add_hashes.finish()
 }
